Fully import Concat, Except, Intersect & Union Tests
Convert Concat, Except, Intersect & Union tests imported in #2874 into a more
directly XUnit-ian style.
