+++ src/System.Linq/tests/System.Linq.Tests.csproj
@@ -21,13 +21,16 @@
     <Compile Include="AllTests.cs" />
     <Compile Include="AnyTests.cs" />
     <Compile Include="CachedEnumerator.cs" />
+    <Compile Include="ConcatTests.cs" />
     <Compile Include="EnumerableTests.cs" />
     <Compile Include="EmptyEnumerable.cs" />
+    <Compile Include="ExceptTests.cs" />
     <Compile Include="FirstOrDefaultTests.cs" />
     <Compile Include="FirstTests.cs" />
     <Compile Include="Helpers\TestCollection.cs" />
     <Compile Include="Helpers\TestEnumerable.cs" />
     <Compile Include="Helpers\TestReadOnlyCollection.cs" />
+    <Compile Include="IntersectTests.cs" />
     <Compile Include="LastOrDefaultTests.cs" />
     <Compile Include="LastTests.cs" />
     <Compile Include="MinMaxTests.cs" />
@@ -51,10 +54,10 @@
     <Compile Include="ToDictionaryTests.cs" />
     <Compile Include="ToListTests.cs" />
     <Compile Include="SelectTests.cs" />
+    <Compile Include="UnionTests.cs" />
     <Compile Include="LegacyTests\AggregateTests.cs" />
     <Compile Include="LegacyTests\AverageTests.cs" />
     <Compile Include="LegacyTests\CastTests.cs" />
-    <Compile Include="LegacyTests\ConcatTests.cs" />
     <Compile Include="LegacyTests\ContainsTests.cs" />
     <Compile Include="LegacyTests\CountTests.cs" />
     <Compile Include="LegacyTests\DefaultIfEmptyTests.cs" />
@@ -62,11 +65,9 @@
     <Compile Include="LegacyTests\ElementAtOrDefaultTests.cs" />
     <Compile Include="LegacyTests\ElementAtTests.cs" />
     <Compile Include="LegacyTests\EqualAllTests.cs" />
-    <Compile Include="LegacyTests\ExceptTests.cs" />
     <Compile Include="LegacyTests\GroupByTests.cs" />
     <Compile Include="LegacyTests\GroupJoinTests.cs" />
     <Compile Include="LegacyTests\Helper.cs" />
-    <Compile Include="LegacyTests\IntersectTests.cs" />
     <Compile Include="LegacyTests\JoinTests.cs" />
     <Compile Include="LegacyTests\LongCountTests.cs" />
     <Compile Include="LegacyTests\MaxTests.cs" />
@@ -91,7 +92,6 @@
     <Compile Include="LegacyTests\ToDictionaryTests.cs" />
     <Compile Include="LegacyTests\ToLookupTests.cs" />
     <Compile Include="LegacyTests\ToSequenceTests.cs" />
-    <Compile Include="LegacyTests\UnionTests.cs" />
     <Compile Include="LegacyTests\WhereTests.cs" />
     <Compile Include="ZipTests.cs" />
   </ItemGroup>
